requirements.txt
scikit-learn
nltk
bs4
selenium
pandas
numpy
matplotlib
python-dotenv
seaborn
googletrans==4.0.0rc1
requests
emoji